Potato integration and MAX_STAT
Made a new variable potato as a currency. This is to be used as currency in the shop and as rewards for completing a task. MAX_STAT has been moved outside main() in order to be used as a global constant. The potato reward for completing/failing a task scales with difficulty.